pyproject.toml
[project]
name = "csm-ft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "librosa>=0.11.0",
    "python-dotenv>=1.1.0",
    "soundfile>=0.13.1",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "transformers~=4.53",
    "typer>=0.16.0",
    "unsloth>=2025.6.4",
    "wandb>=0.20.1",
]


[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true


[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
xformers = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
